app-2/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.27,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pandas>=2.2,<3.0
pydantic
python-dotenv
scalar-fastapi==1.8.0